+++ controllers/messagesController.js
@@ -4,13 +4,19 @@ const { Op } = require("sequelize")
 
 
 class MessagesController {
-    async create(req, res) {
+    async create(req, res, next) {
         const {message, outgoing_id, incoming_id} = req.body;
+        if(!message || !outgoing_id || !incoming_id) {
+            return next(ApiError.badRequest('Не задано одно из обязательных полей'));
+        }
         const msg = await Messages.create({message, outgoing_id, incoming_id});
         return res.json(msg);
     }
-    async getAll(req, res) {
+    async getAll(req, res, next) {
         let {firstUserId, secondUserId, limit, page} = req.query;
+        if(!firstUserId || !secondUserId) {
+            return next(ApiError.badRequest('Не задано одно из обязательных полей'));
+        }
         // page = page || 1;
         // limit = limit || 30;
         // let offset = page * limit - limit;
@@ -32,21 +38,30 @@ class MessagesController {
     });
         return res.json(messages);
     }
-    async findMessages(req, res) {
+    async findMessages(req, res, next) {
         let {queryParameter, id, page, limit} = req.query;
+        if(!queryParameter || !id) {
+            return next(ApiError.badRequest('Не задано одно из обязательных полей'));
+        }
         page = page || 1;
         limit = limit || 20;
         let offset = page * limit - limit;
         const msg = await Messages.findAll({where: {message: {[Op.substring]: queryParameter}, [Op.or]: [{outgoing_id: id}, {incoming_id:id}]}}, limit)
         return res.json(msg);
     }
-    async updateMessage(req, res) {
+    async updateMessage(req, res, next) {
         const {id, message} = req.body;
+        if(!message || !id) {
+            return next(ApiError.badRequest('Не задано одно из обязательных полей'));
+        }
         const updatedMessage = await Messages.update({message}, {where: {id: id}})
         return res.json(updatedMessage)
     }
     async deleteMessage(req, res) {
         const {id} = req.body;
+        if(!id) {
+            return next(ApiError.badRequest('Не задано обязательное поле'));
+        }
         const deletedMessage = await Messages.destroy({where: {id: id}})
         return res.json(deletedMessage)
     }
